multiple updates.
with this update, i have added a royalty free image as a background. This will be grown in the future.
I have also added space to the GUI for future development.
i have removed the original weather-app.py to avoid confusion when downloading it.
I have also begun developing lists for the response to make it a bit more fun and engaging.
all of the changes above are geared towards more growth within this app